Refactor(#38): package structure 수정 및 리팩토링
